requirements.txt
torch
transformers
gradio
diskcache
datasets
tqdm
numpy
Pillow
torchvision
torchmeta
higher
